Fix printf of sa_family_t variables.
sa_family_t is different sizes on different platforms (eg unsigned char on osx;
unsigned short on linux).  For portability, cast to uintmax_t, since we're
guaranteed to avoid loss of precision, and then use the format specifier for
that type.

+++ tests/common/sockaddrtest.c
@@ -1,13 +1,16 @@
 #ifndef GUARD_TESTS_SOCKADDRTEST_H
 #define GUARD_TESTS_SOCKADDRTEST_H
 
+#include <inttypes.h>
+
 #include "config.h"
 
 #include "ntp.h"
 
 bool IsEqualS(const sockaddr_u *expected, const sockaddr_u *actual) {
 	if (expected->sa.sa_family != actual->sa.sa_family) {
-		printf("Expected sa_family: %hhu but got: %hhu\n", expected->sa.sa_family, actual->sa.sa_family);
+		printf("Expected sa_family: %" PRIuMAX " but got: %" PRIuMAX "\n",
+		       (uintmax_t)expected->sa.sa_family, (uintmax_t)actual->sa.sa_family);
 		return false;
 	}
 
@@ -30,7 +33,7 @@ bool IsEqualS(const sockaddr_u *expected, const sockaddr_u *actual) {
 			return false;
 		}
 	} else { // Unknown family
-		printf("Unknown sa_family: %hhu\n", actual->sa.sa_family);
+		printf("Unknown sa_family: %" PRIuMAX "\n", (uintmax_t)actual->sa.sa_family);
 		return false;
 	}
 }
